Refact BoardsClient to ProjectClient
